Cleaned up pr after review.
