+++ tests/src/Functional/Lock/RedisLockFunctionalTest.php
@@ -5,6 +5,7 @@ namespace Drupal\Tests\redis\Functional\Lock;
 use Drupal\Component\Utility\OpCodeCache;
 use Drupal\Core\Site\Settings;
 use Drupal\Tests\system\Functional\Lock\LockFunctionalTest;
+use Drupal\Tests\redis\Traits\RedisTestInterfaceTrait;
 
 /**
  * Confirm locking works between two separate requests.
@@ -12,6 +13,8 @@ use Drupal\Tests\system\Functional\Lock\LockFunctionalTest;
  * @group redis
  */
 class RedisLockFunctionalTest extends LockFunctionalTest {
+
+  use RedisTestInterfaceTrait;
 
   /**
    * Modules to enable.
@@ -31,7 +34,7 @@ class RedisLockFunctionalTest extends LockFunctionalTest {
     $filename = $this->siteDirectory . '/settings.php';
     chmod($filename, 0666);
     $contents = file_get_contents($filename);
-    $redis_interface = getenv('REDIS_INTERFACE');
+    $redis_interface = self::getRedisInterfaceEnv();
     $contents .= "\n\n" . '$settings[\'container_yamls\'][] = \'modules/redis/example.services.yml\';';
     $contents .= "\n\n" . '$settings["redis.connection"]["interface"] = \'' . $redis_interface . '\';';
     file_put_contents($filename, $contents);
