+++ tests/src/Kernel/RedisFloodTest.php
@@ -4,7 +4,7 @@ namespace Drupal\Tests\redis\Kernel;
 
 use Drupal\KernelTests\KernelTestBase;
 use Drupal\redis\Flood\PhpRedis;
-use Drupal\Core\Site\Settings;
+use Drupal\Tests\redis\Traits\RedisTestInterfaceTrait;
 
 /**
  * Tests Redis flood backend.
@@ -12,6 +12,8 @@ use Drupal\Core\Site\Settings;
  * @group redis
  */
 class RedisFloodTest extends KernelTestBase {
+
+  use RedisTestInterfaceTrait;
 
   /**
    * Modules to enable.
@@ -51,18 +53,6 @@ class RedisFloodTest extends KernelTestBase {
     $_SERVER['REQUEST_TIME'] += 2;
     $this->assertTrue($flood->isAllowed($name, $threshold));
 
-  }
-
-  /**
-   * Uses an env variable to set the redis client to use for this test.
-   */
-  protected function setUpSettings() {
-
-    // Write redis_interface settings manually.
-    $redis_interface = getenv('REDIS_INTERFACE');
-    $settings = Settings::getAll();
-    $settings['redis.connection']['interface'] = $redis_interface;
-    new Settings($settings);
   }
 
 }
